Refactor contact command into CRUD
