+++ src/main/java/seedu/address/logic/commands/update/RemoveTagFromContactCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.contact;
+package seedu.address.logic.commands.update;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.logic.parser.CliSyntax.CONTACT_COMMAND_WORD;
@@ -9,30 +9,29 @@ import java.util.Set;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.Messages;
-import seedu.address.logic.commands.update.EditCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.contact.Contact;
 import seedu.address.model.contact.Tag;
 
 /**
- * Adds a tag to a Contact in contact.
+ * Removes a tag from a specified contact.
  */
-public class AddTagToContactCommand extends EditCommand<Contact> {
-    public static final String COMMAND_WORD = "tag";
+public class RemoveTagFromContactCommand extends EditCommand<Contact> {
+    public static final String COMMAND_WORD = "untag";
     public static final String MESSAGE_USAGE = CONTACT_COMMAND_WORD + " " + COMMAND_WORD
-            + ": Adds a tag to a specified contact.\n"
+            + ": Removes a tag from a specified contact.\n"
             + "Parameters: INDEX " + PREFIX_CONTACT_TAG_LONG + " <tag>\n"
             + "Example: " + CONTACT_COMMAND_WORD + " " + COMMAND_WORD + " 1 "
             + PREFIX_CONTACT_TAG_LONG + " TA";
-    public static final String MESSAGE_ADD_TAG_SUCCESSFUL = "Added tag to contact: %1$s";
-    public static final String MESSAGE_DUPLICATE_TAGS = "The tag is already assigned to this contact";
+    public static final String MESSAGE_REMOVE_TAG_SUCCESSFUL = "Removed tag from contact: %1$s";
+    public static final String MESSAGE_MISSING_TAG = "The tag is already removed from this contact";
     private final Tag tag;
 
     /**
-     * Creates an EditCommand to add a tag to the specified {@code Contact}
+     * Creates an EditCommand to remove a tag from a specified {@code Contact}.
      */
-    public AddTagToContactCommand(Index index, Tag tag) {
+    public RemoveTagFromContactCommand(Index index, Tag tag) {
         super(index, Model::getContactManagerAndList);
         requireNonNull(tag);
         this.tag = tag;
@@ -40,18 +39,19 @@ public class AddTagToContactCommand extends EditCommand<Contact> {
 
     @Override
     public Contact createEditedItem(Model model, Contact contactToEdit) throws CommandException {
-        if (contactToEdit.getTags().contains(tag)) {
-            throw new CommandException(String.format(MESSAGE_DUPLICATE_TAGS));
+        if (!contactToEdit.getTags().contains(tag)) {
+            throw new CommandException(MESSAGE_MISSING_TAG);
         }
         Set<Tag> newTags = new HashSet<>(contactToEdit.getTags());
-        newTags.add(tag);
+        newTags.remove(tag);
         return new Contact(
-                contactToEdit.getId(),
-                contactToEdit.getName(),
-                contactToEdit.getEmail(),
-                contactToEdit.getCourse(),
-                contactToEdit.getGroup(),
-                Set.copyOf(newTags));
+            contactToEdit.getId(),
+            contactToEdit.getName(),
+            contactToEdit.getEmail(),
+            contactToEdit.getCourse(),
+            contactToEdit.getGroup(),
+            Set.copyOf(newTags)
+        );
     }
 
     @Override
@@ -61,11 +61,11 @@ public class AddTagToContactCommand extends EditCommand<Contact> {
 
     @Override
     public String getDuplicateMessage() {
-        return MESSAGE_DUPLICATE_TAGS;
+        return Messages.MESSAGE_DUPLICATE_FIELDS;
     }
 
     @Override
     public String getSuccessMessage(Contact editedContact) {
-        return String.format(MESSAGE_ADD_TAG_SUCCESSFUL, Messages.format(editedContact));
+        return String.format(MESSAGE_REMOVE_TAG_SUCCESSFUL, Messages.format(editedContact));
     }
 }
