+++ src/main/java/seedu/address/logic/parser/contact/AddTagToContactCommandParser.java
@@ -5,7 +5,7 @@ import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.contact.ContactCliSyntax.PREFIX_CONTACT_TAG_LONG;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.contact.AddTagToContactCommand;
+import seedu.address.logic.commands.update.AddTagToContactCommand;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
