+++ src/main/java/seedu/address/logic/parser/contact/FilterContactCommandParser.java
@@ -21,7 +21,7 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import seedu.address.commons.core.Operator;
-import seedu.address.logic.commands.contact.FilterContactCommand;
+import seedu.address.logic.commands.read.FilterContactCommand;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
