+++ src/main/java/seedu/address/ui/MainWindow.java
@@ -14,7 +14,7 @@ import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.logic.Logic;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.contact.ListContactCommand;
+import seedu.address.logic.commands.read.ListContactCommand;
 import seedu.address.logic.commands.event.ListEventCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.commands.todo.ListTodoCommand;
