+++ src/test/java/seedu/address/logic/commands/CommandTestUtil.java
@@ -14,7 +14,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.contact.EditContactCommand;
+import seedu.address.logic.commands.update.EditContactCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.contact.Contact;
