+++ src/test/java/seedu/address/logic/commands/contact/FilterContactCommandTest.java
@@ -18,6 +18,7 @@ import java.util.Map;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.Operator;
+import seedu.address.logic.commands.read.FilterContactCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
