+++ src/test/java/seedu/address/logic/parser/HelpCommandParserTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.HelpCommand;
-import seedu.address.logic.commands.contact.AddContactCommand;
+import seedu.address.logic.commands.create.AddContactCommand;
 import seedu.address.logic.commands.event.AddEventCommand;
 import seedu.address.logic.commands.todo.AddTodoCommand;
 import seedu.address.model.Model;
