+++ src/test/java/seedu/address/logic/parser/contact/DeleteContactCommandParserTest.java
@@ -7,7 +7,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.contact.DeleteContactCommand;
+import seedu.address.logic.commands.delete.DeleteContactCommand;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path variations
