+++ src/test/java/seedu/address/logic/parser/contact/FilterContactCommandParserTest.java
@@ -17,7 +17,7 @@ import java.util.Map;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.Operator;
-import seedu.address.logic.commands.contact.FilterContactCommand;
+import seedu.address.logic.commands.read.FilterContactCommand;
 import seedu.address.model.contact.ContactColumn;
 import seedu.address.model.contact.ContactPredicate;
 import seedu.address.model.contact.FilterCriteria;
